v3.0.2-rc.1: deprecate nongeneric CompareTo
