+++ src/core-unit/Unit.Tests.Old/UnitTests/UnitTests.Comparison.cs
@@ -16,38 +16,6 @@ partial class UnitTests
     }
 
     [Test]
-    public void CompareTo_Obj_Unit_ExpectZero()
-    {
-        object? obj = default(Unit);
-
-        var actual = Unit.Value.CompareTo(obj);
-        Assert.That(actual, Is.Zero);
-    }
-
-    [Test]
-    public void CompareTo_Obj_Null_ExpectGreaterThanZero_ExpectOne()
-    {
-        object? obj = null;
-
-        var actual = Unit.Value.CompareTo(obj);
-
-        Assert.That(actual, Is.Positive);
-        Assert.That(actual, Is.EqualTo(1));
-    }
-
-    [Test]
-    [TestCaseSource(nameof(CompareTo_Obj_NotUnit_ExpectArgumentException_TestCases))]
-    public void CompareTo_Obj_NotUnit_ExpectArgumentException(object obj)
-    {
-        var ex = Assert.Throws<ArgumentException>(() => _ = Unit.Value.CompareTo(obj))!;
-
-        Assert.That(ex.ParamName, Is.EqualTo("obj"));
-
-        var containsExpectedMessage = ex.Message.Contains("The object is not Unit.", StringComparison.Ordinal);
-        Assert.That(containsExpectedMessage, Is.True);
-    }
-
-    [Test]
     public void Compare_Static_ExpectZero()
     {
         var left = Unit.Value;
@@ -95,14 +63,5 @@ partial class UnitTests
 
         var actual = left > right;
         Assert.That(actual, Is.False);
-    }
-
-    private static IEnumerable<object> CompareTo_Obj_NotUnit_ExpectArgumentException_TestCases()
-    {
-        yield return new object();
-        yield return PlusFifteen;
-        yield return PlusFifteenIdRefType;
-        yield return SomeTextStructType;
-        yield return Array.Empty<Unit>();
     }
 }
