+++ src/core-unit/Unit/Unit.csproj
@@ -15,11 +15,11 @@
     <RepositoryUrl>https://github.com/pfpack/pfpack-core-unit</RepositoryUrl>
     <Company>pfpack</Company>
     <Authors>Andrei Sergeev, Pavel Moskovoy</Authors>
-    <Copyright>Copyright © 2020-2024 Andrei Sergeev, Pavel Moskovoy</Copyright>
+    <Copyright>Copyright © 2020-2025 Andrei Sergeev, Pavel Moskovoy</Copyright>
     <Description>PrimeFuncPack Core.Unit is a core library for .NET consisting of Unit type targeted for use in functional programming.</Description>
     <RootNamespace>System</RootNamespace>
     <AssemblyName>PrimeFuncPack.Core.Unit</AssemblyName>
-    <Version>3.0.1</Version>
+    <Version>3.0.2-rc.1</Version>
   </PropertyGroup>
 
   <ItemGroup>
