+++ src/core-unit/Unit/Unit/Unit.Comparison.cs
@@ -3,13 +3,4 @@
 partial struct Unit
 {
     public int CompareTo(Unit other) => default;
-
-    public int CompareTo(object? obj) => obj switch
-    {
-        null => 1,
-
-        Unit => default,
-
-        _ => throw new ArgumentException("The object is not Unit.", nameof(obj))
-    };
 }
